lng_rl_optimizer/pyproject.toml
[project]
name = "lng_rl_optimizer"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "CoolProp>=6.6",
    "torch==2.1.0",
    "stable-baselines3==2.2.1",
    "gymnasium==0.29.1",
    "pandas>=2.1",
    "numpy>=1.26",
    "scipy>=1.11",
    "matplotlib>=3.8",
    "mlflow>=2.8",
    "tqdm>=4.66",
    "click>=8.1",
    "pyyaml>=6.0",
    "requests>=2.31",
    "scikit-learn>=1.3",
]
[project.optional-dependencies]
dev = ["pytest>=7.4", "jupyter>=1.0"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
